pyproject.toml
[tool.poetry]
name = "loglens"
version = "0.1.0"
description = "LogLens - Advanced Log Analysis Tool for security monitoring and operational insights"
authors = ["LogLens Contributors <contributors@loglens.dev>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/your-org/loglens"
repository = "https://github.com/your-org/loglens"
documentation = "https://github.com/your-org/loglens/blob/main/README.md"
keywords = ["log-analysis", "security", "monitoring", "apache-logs", "pattern-detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
    "Topic :: Security",
]
packages = [{include = "loglens", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pydantic = "^2.0.0"
httpx = "^0.25.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
loglens = "loglens.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 